[PLAT-569] flatten models folder
